backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
drf-yasg>=1.21
psycopg2-binary>=2.9
redis>=5.0
django-redis>=5.4
python-dotenv>=1.0
gunicorn>=21.2
Pillow>=10.0
django-filter>=23.0
djangorestframework-simplejwt>=5.3
pytest-django>=4.7
pytest>=7.4
pytest-cov>=4.1
factory-boy>=3.3
